labeleR: Automate the Production of Custom Labels, Badges, Certificates, and Other Documents

Create custom labels, badges, certificates and other documents. Automate the production of potentially large numbers of accreditation badges, attendance and participation certificates, herbarium and collection labels, etc. Documents are generated in PDF format, which requires a working installation of 'LaTeX', such as 'TinyTeX'.

See also

Author

Maintainer: Ignacio Ramos-Gutierrez ig.ramosgutierrez@gmail.com (ORCID) [copyright holder]

Authors:

Julia G. de Aledo juliagdealedo@gmail.com (ORCID) [copyright holder]

Francisco Rodriguez-Sanchez frodriguez@us.es (ORCID) [copyright holder]

Other contributors:

Giorgio Comai [contributor]