tiny.table

Example data to create tinylabels

Usage

tiny.table

Format

An object of class data.frame with 40 rows and 8 columns.

Details

field1	field2	field3	field4	field5	field6	field7	QR_code

Bombus terrestris subsp. glumbumble	Apidae	05/09/1992	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Item 1	Collected in the coast of the North Sea (it was only known from the Mediterranean Sea before).	It produces a fluid that causes melancholy. This fluid is used as an antidote for the hysteria that results from eating Alihotsy leaves.	https://static.wikia.nocookie.net/esharrypotter/images/b/b2/HM_Glumblumble.png/revision/latest?cb=20210127125033

Bombus terrestris subsp. glumbumble	Apidae	01/09/1992	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Item 2	Collected in the coast of the North Sea (it was only known from the Mediterranean Sea before).	It produces a fluid that causes melancholy. This fluid is used as an antidote for the hysteria that results from eating Alihotsy leaves.	https://static.wikia.nocookie.net/esharrypotter/images/b/b2/HM_Glumblumble.png/revision/latest?cb=20210127125033

Bombus terrestris subsp. glumbumble	Apidae	13/11/1992	Cargan, Antrim Glens, Northern Ireland.	Item 3	Collected in the coast of the North Sea (it was only known from the Mediterranean Sea before).	It produces a fluid that causes melancholy. This fluid is used as an antidote for the hysteria that results from eating Alihotsy leaves.	https://static.wikia.nocookie.net/esharrypotter/images/b/b2/HM_Glumblumble.png/revision/latest?cb=20210127125033

Bombus terrestris subsp. glumbumble	Apidae	13/12/1992	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Item 4	Collected in the coast of the North Sea (it was only known from the Mediterranean Sea before).	It produces a fluid that causes melancholy. This fluid is used as an antidote for the hysteria that results from eating Alihotsy leaves.	https://static.wikia.nocookie.net/esharrypotter/images/b/b2/HM_Glumblumble.png/revision/latest?cb=20210127125033

Amata huebneri	Erebidae	05/09/1992	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Item 1	Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm.	Wasp-moth hybrid, giant and aggresive, painful sting	https://harrypotter.fandom.com/wiki/Vampyr_Mosp?file=Vampyr_Mosp.png

Amata huebneri	Erebidae	01/09/1992	Great Yarmouth, Norwich, United Kingdom	Item 2	Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm.	Wasp-moth hybrid, giant and aggresive, painful sting	https://harrypotter.fandom.com/wiki/Vampyr_Mosp?file=Vampyr_Mosp.png

Amata huebneri	Erebidae	13/11/1992	Cargan, Antrim Glens, Northern Ireland.	Item 3	Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm.	Wasp-moth hybrid, giant and aggresive, painful sting	https://harrypotter.fandom.com/wiki/Vampyr_Mosp?file=Vampyr_Mosp.png

Amata huebneri	Erebidae	13/12/1992	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Item 4	Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm.	Wasp-moth hybrid, giant and aggresive, painful sting	https://harrypotter.fandom.com/wiki/Vampyr_Mosp?file=Vampyr_Mosp.png

Arion hortensis	Arionidae	05/09/1992	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Item 1	Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm.	Ron Weasley began vomiting slugs after his attempt to curse Draco Malfoy failed due to his broken wand.	https://figuritaspop.es/wp-content/uploads/ronweasleyslug1_2.jpg

Arion hortensis	Arionidae	01/09/1992	Great Yarmouth, Norwich, United Kingdom	Item 2	Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm.	Ron Weasley began vomiting slugs after his attempt to curse Draco Malfoy failed due to his broken wand.	https://figuritaspop.es/wp-content/uploads/ronweasleyslug1_2.jpg

Arion hortensis	Arionidae	13/11/1992	Cargan, Antrim Glens, Northern Ireland.	Item 3	Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm.	Ron Weasley began vomiting slugs after his attempt to curse Draco Malfoy failed due to his broken wand.	https://figuritaspop.es/wp-content/uploads/ronweasleyslug1_2.jpg

Arion hortensis	Arionidae	13/12/1992	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Item 4	Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm.	Ron Weasley began vomiting slugs after his attempt to curse Draco Malfoy failed due to his broken wand.	https://figuritaspop.es/wp-content/uploads/ronweasleyslug1_2.jpg

Acromantula gigantea	Macroaranidae	05/09/1992	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Item 1	Forbidden Forest	small individuals of approx. 30 cm have been collected.	https://mymagicwands.files.wordpress.com/2020/04/img_20200409_115122.jpg

Acromantula gigantea	Macroaranidae	01/09/1992	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Item 2	Forbidden Forest	small individuals of approx. 30 cm have been collected.	https://mymagicwands.files.wordpress.com/2020/04/img_20200409_115122.jpg

Acromantula gigantea	Macroaranidae	13/11/1992	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Item 3	Forbidden Forest	small individuals of approx. 30 cm have been collected.	https://mymagicwands.files.wordpress.com/2020/04/img_20200409_115122.jpg

Acromantula gigantea	Macroaranidae	13/12/1992	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Item 4	Forbidden Forest	small individuals of approx. 30 cm have been collected.	https://mymagicwands.files.wordpress.com/2020/04/img_20200409_115122.jpg

Trorgssionis sp. 1	Canellae	10/5/23	Grazalema mountain range (Cadiz, Spain)	Item 1	Between the bushes of Retama and Nerium	Blue colour of the skin indicates adaptation to the field conditions	https://www.androidheadlines.com/wp-content/uploads/2021/11/google-search-fishing-troll.jpg

Trorgssionis sp. 2	Canellae	10/5/23	Grazalema mountain range (Cadiz, Spain)	Item 2	Between the bushes of Retama and Nerium	Blue colour of the skin indicates adaptation to the field conditions	https://www.androidheadlines.com/wp-content/uploads/2021/11/google-search-fishing-troll.jpg

Trorgssionis schetelonica	Canellae	10/5/23	Grazalema mountain range (Cadiz, Spain)	Item 1	Between the bushes of Retama and Nerium	Blue colour of the skin indicates adaptation to the field conditions	https://www.androidheadlines.com/wp-content/uploads/2021/11/google-search-fishing-troll.jpg

Trorgssionis schetelonica	Canellae	10/5/23	Grazalema mountain range (Cadiz, Spain)	Item 2	Between the bushes of Retama and Nerium	Blue colour of the skin indicates adaptation to the field conditions	https://www.androidheadlines.com/wp-content/uploads/2021/11/google-search-fishing-troll.jpg

Bombus terrestris subsp. glumbumble	Apidae	05/09/1992	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Item 5	Collected in the coast of the North Sea (it was only known from the Mediterranean Sea before).	It produces a fluid that causes melancholy. This fluid is used as an antidote for the hysteria that results from eating Alihotsy leaves.	https://static.wikia.nocookie.net/esharrypotter/images/b/b2/HM_Glumblumble.png/revision/latest?cb=20210127125033

Bombus terrestris subsp. glumbumble	Apidae	01/09/1992	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Item 6	Collected in the coast of the North Sea (it was only known from the Mediterranean Sea before).	It produces a fluid that causes melancholy. This fluid is used as an antidote for the hysteria that results from eating Alihotsy leaves.	https://static.wikia.nocookie.net/esharrypotter/images/b/b2/HM_Glumblumble.png/revision/latest?cb=20210127125033

Bombus terrestris subsp. glumbumble	Apidae	13/11/1992	Cargan, Antrim Glens, Northern Ireland.	Item 7	Collected in the coast of the North Sea (it was only known from the Mediterranean Sea before).	It produces a fluid that causes melancholy. This fluid is used as an antidote for the hysteria that results from eating Alihotsy leaves.	https://static.wikia.nocookie.net/esharrypotter/images/b/b2/HM_Glumblumble.png/revision/latest?cb=20210127125033

Bombus terrestris subsp. glumbumble	Apidae	13/12/1992	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Item 8	Collected in the coast of the North Sea (it was only known from the Mediterranean Sea before).	It produces a fluid that causes melancholy. This fluid is used as an antidote for the hysteria that results from eating Alihotsy leaves.	https://static.wikia.nocookie.net/esharrypotter/images/b/b2/HM_Glumblumble.png/revision/latest?cb=20210127125033

Amata huebneri	Erebidae	05/09/1992	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Item 5	Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm.	Wasp-moth hybrid, giant and aggresive, painful sting	https://harrypotter.fandom.com/wiki/Vampyr_Mosp?file=Vampyr_Mosp.png

Amata huebneri	Erebidae	01/09/1992	Great Yarmouth, Norwich, United Kingdom	Item 6	Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm.	Wasp-moth hybrid, giant and aggresive, painful sting	https://harrypotter.fandom.com/wiki/Vampyr_Mosp?file=Vampyr_Mosp.png

Amata huebneri	Erebidae	13/11/1992	Cargan, Antrim Glens, Northern Ireland.	Item 7	Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm.	Wasp-moth hybrid, giant and aggresive, painful sting	https://harrypotter.fandom.com/wiki/Vampyr_Mosp?file=Vampyr_Mosp.png

Amata huebneri	Erebidae	13/12/1992	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Item 8	Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm.	Wasp-moth hybrid, giant and aggresive, painful sting	https://harrypotter.fandom.com/wiki/Vampyr_Mosp?file=Vampyr_Mosp.png

Arion hortensis	Arionidae	05/09/1992	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Item 5	Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm.	Ron Weasley began vomiting slugs after his attempt to curse Draco Malfoy failed due to his broken wand.	https://figuritaspop.es/wp-content/uploads/ronweasleyslug1_2.jpg

Arion hortensis	Arionidae	01/09/1992	Great Yarmouth, Norwich, United Kingdom	Item 6	Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm.	Ron Weasley began vomiting slugs after his attempt to curse Draco Malfoy failed due to his broken wand.	https://figuritaspop.es/wp-content/uploads/ronweasleyslug1_2.jpg

Arion hortensis	Arionidae	13/11/1992	Cargan, Antrim Glens, Northern Ireland.	Item 7	Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm.	Ron Weasley began vomiting slugs after his attempt to curse Draco Malfoy failed due to his broken wand.	https://figuritaspop.es/wp-content/uploads/ronweasleyslug1_2.jpg

Arion hortensis	Arionidae	13/12/1992	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Item 8	Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm.	Ron Weasley began vomiting slugs after his attempt to curse Draco Malfoy failed due to his broken wand.	https://figuritaspop.es/wp-content/uploads/ronweasleyslug1_2.jpg

Acromantula gigantea	Macroaranidae	05/09/1992	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Item 5	Forbidden Forest	small individuals of approx. 30 cm have been collected.	https://mymagicwands.files.wordpress.com/2020/04/img_20200409_115122.jpg

Acromantula gigantea	Macroaranidae	01/09/1992	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Item 6	Forbidden Forest	small individuals of approx. 30 cm have been collected.	https://mymagicwands.files.wordpress.com/2020/04/img_20200409_115122.jpg

Acromantula gigantea	Macroaranidae	13/11/1992	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Item 7	Forbidden Forest	small individuals of approx. 30 cm have been collected.	https://mymagicwands.files.wordpress.com/2020/04/img_20200409_115122.jpg

Acromantula gigantea	Macroaranidae	13/12/1992	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Item 8	Forbidden Forest	small individuals of approx. 30 cm have been collected.	https://mymagicwands.files.wordpress.com/2020/04/img_20200409_115122.jpg

Trorgssionis sp. 1	Canellae	10/5/23	Grazalema mountain range (Cadiz, Spain)	Item 5	Between the bushes of Retama and Nerium	Blue colour of the skin indicates adaptation to the field conditions	https://www.androidheadlines.com/wp-content/uploads/2021/11/google-search-fishing-troll.jpg

Trorgssionis sp. 2	Canellae	10/5/23	Grazalema mountain range (Cadiz, Spain)	Item 6	Between the bushes of Retama and Nerium	Blue colour of the skin indicates adaptation to the field conditions	https://www.androidheadlines.com/wp-content/uploads/2021/11/google-search-fishing-troll.jpg

Trorgssionis schetelonica	Canellae	10/5/23	Grazalema mountain range (Cadiz, Spain)	Item 7	Between the bushes of Retama and Nerium	Blue colour of the skin indicates adaptation to the field conditions	https://www.androidheadlines.com/wp-content/uploads/2021/11/google-search-fishing-troll.jpg

Trorgssionis schetelonica	Canellae	10/5/23	Grazalema mountain range (Cadiz, Spain)	Item 8	Between the bushes of Retama and Nerium	Blue colour of the skin indicates adaptation to the field conditions	https://www.androidheadlines.com/wp-content/uploads/2021/11/google-search-fishing-troll.jpg