Function to create herbaruim labels (4 per DIN-A4 page)

Function to create herbaruim labels (4 per DIN-A4 page)

Usage

create_herbarium_label(
  data = data,
  path = NULL,
  title = NULL,
  subtitle = NULL,
  qr = NULL,
  family.column = NULL,
  taxon.column = NULL,
  author.column = NULL,
  det.column = NULL,
  date.det.column = NULL,
  location.column = NULL,
  area.description.column = NULL,
  latitude.column = NULL,
  longitude.column = NULL,
  elevation.column = NULL,
  field1.column = NULL,
  field2.column = NULL,
  field3.column = NULL,
  collector.column = NULL,
  collection.column = NULL,
  assistants.column = NULL,
  date.column = NULL
)

Arguments

data

a data frame to create herbarium labels.

path

Folder path where the output will be printed

title

Main title at the top of the labels. Can be blank if set to NULL.

subtitle

Subtitle at the bottom of the labels. Can be blank if set to NULL.

qr

String. Free text or column of data that specifies the text to create the QR code. If the specified value is not a column name of data, all the QRs will be equal, and will output the specified qr.

family.column

Column name of the data data frame which specifies the labels' family name.

taxon.column

Column name of the data data frame which specifies the labels' taxon.

author.column

Column name of the data data frame which specifies the taxons' author.

det.column

Column name of the data data frame which specifies the determiner of the voucher.

date.det.column

Column name of the data data frame which specifies the date when the voucher was determined.

location.column

Column name of the data data frame which specifies where the voucher was collected.

area.description.column

Column name of the data data frame which specifies the decription of the area

latitude.column

Column name of the data data frame which specifies the latitude where the specimen was collected.

longitude.column

Column name of the data data frame which specifies the longitude where the specimen was collected.

elevation.column

Column name of the data data frame which specifies the elevation where the specimen was collected.

field1.column

Column name of the data data frame which specifying a variable of the user's choice. Can be blank if set to NULL.

field2.column

Column name of the data data frame which specifying a variable of the user's choice. Can be blank if set to NULL.

field3.column

Column name of the data data frame which specifying a variable of the user's choice. Can be blank if set to NULL.

collector.column

Column name of the data data frame which specifies the name of the collector of the voucher.

collection.column

Column name of the data data frame which specifies the voucher's collection number.

assistants.column

Column name of the data data frame which specifies the names of the collector's assistants.

date.column

Column name of the data data frame which specifies the date when the specimen was collected.

Value

A pdf file with four herbarium labels per page within an 'output' folder

Author

Julia G. de Aledo, Ignacio Ramos-Gutierrez

Examples

#
data=read_sheet("https://docs.google.com/spreadsheets/
        d/1Q005BDM0XyUNq5XzGWuvdzgZVMc4KhbYadVzi77h3Xw/edit?usp=sharing")
#> Error in open.connection(structure(8L, class = c("curl", "connection"), conn_id = <pointer: 0x00000000000004d8>),     "rb"): Timeout was reached: [doc-0s-4o-sheets.googleusercontent.com] Operation timed out after 10004 milliseconds with 0 out of 0 bytes received
create_herbarium_label(
data=data,
path = "LabeleR_output",
title="Magical flora of the British Isles",
subtitle="Project: Eliminating plant blindness in Hogwarts students",
qr = "QR_code",
family.column="Family",
taxon.column="Taxon",
author.column="Author",
det.column="det/conf",
date.det.column="Det_date",
location.column="Location",
area.description.column="Area_description",
latitude.column="Latitude",
longitude.column="Longitude",
elevation.column="Elevation",
field1.column="life_form",
field2.column="Observations",
field3.column="Height",
collector.column="Collector",
collection.column="Collection_number",
assistants.column="Assistants",
date.column="Date"
)
#> Error in create_herbarium_label(data = data, path = "LabeleR_output",     title = "Magical flora of the British Isles", subtitle = "Project: Eliminating plant blindness in Hogwarts students",     qr = "QR_code", family.column = "Family", taxon.column = "Taxon",     author.column = "Author", det.column = "det/conf", date.det.column = "Det_date",     location.column = "Location", area.description.column = "Area_description",     latitude.column = "Latitude", longitude.column = "Longitude",     elevation.column = "Elevation", field1.column = "life_form",     field2.column = "Observations", field3.column = "Height",     collector.column = "Collector", collection.column = "Collection_number",     assistants.column = "Assistants", date.column = "Date"): The 'data' object must be a data frame.