herbarium.table

Example data to create herbarium labels

herbarium.table

Format

An object of class data.frame with 6 rows and 19 columns.

Details

ID	Collector	Collection_number	Assistants	Family	Taxon	Author	det	Det_date	life_form	Observations	Height	Location	Area_description	Elevation	Date	Latitude	Longitude	QR_code

1	Hermione Granger	2	Sprout, P., Potter, H., Weasley, R.	SOLANACEAE	Mandragora officinarum	L.	Sprout, P.	05/09/1995	Herb	Collected at the seedling stage; loud and harmful crying noise	45cm. Root diameter 2.5cm aprox.	Cargan, Antrim Glens, Northern Ireland.	Marsh vegetation; mainly dominated by sedges and rushes.	12 m.a.s.l.	25/07/94	54.9886 N	-6.1392 E	https://www.clarin.com/img/2021/12/20/5jTLngUq9_1256x620__1.jpg

2	Ron Weasley	5	Potter, H.	SALICACEAE	Salix muhammadalii subsp. robusta	Huff.	Sprout, P.	01/09/1992	Tree	Only known extant individual, near to the Hogwarts Castle.	7.5 m.	Hogwarts Gardens, Hogsmeade, Scotland (UK)	Open grassland, near the Foridden Forest.	250 m.a.s.l.	01/09/92	57.4223 N	-3.0869 E	https://steamuserimages-a.akamaihd.net/ugc/1483325459410929390/8FE97A548651369A35C35018271913794CAD7CEA/?imw=512&&ima=fit&impolicy=Letterbox&imcolor=%23000000&letterbox=false

3	Dobby H.E.	1	Winky H.E.	Zosteraceae	Zostera phycobranchia	(Huff.) Sprout	Longbottom, N.	13/11/1994	Aquatic herb	First known locality outside the Mediterranean Sea.	Maximum height of 15 cm, but rhizomes deep to 40-60 cm into the ground.	Great Yarmouth, Norwich, United Kingdom	15 m from the shore into the sea, forming a monospecific subacuatic prairie.	-3 m.a.s.l.	10/11/1994	52.629060 N	1.745356 E	http://harrypotterdiagon.blogspot.com/2018/01/branquialgas.html

4	Pomona Sprout	1990-301a	Dumbledore, A.	Suffocaceae	Lucifericaptia obscura	Slyth.	Sprout, P.	01/07/1990	Bush	Collected 15m beneath the ground level, where no sunlight could reach.	1.5 m.	South Stoke, Arundel, United Kingdom.	Karstic cave. No other plants found due to the lack of light.	25 m.a.s.l.	08/05/86	50.8638 N	-0.5599 E	https://pbs.twimg.com/media/DbpafnCWAAEOI-X.jpg

5	Luna Lovegood	5	Xenophilius Lovegood	rosaceae	Prunus zeppelin	Rav.	Lovegood, X.	06/08/1996	Tree	Cultivated for gardening purposes.	2 m.	Ottery St. Catchpole, Devon, England (UK).	Garden	40 m.a.s.l.	06/08/1996	50.7602 N	-3.2653 E	https://static.wikia.nocookie.net/habbo-potter/images/8/85/Dirigible_Plum.jpg/revision/latest?cb=20180611174816

6	Algie Longbottom	3	Longbottom, N.	Mimbulaceae	Mimbulus mimbletonia	Spore	Sprout, P.	25/10/1995	Herb	Desert plant usually used as ornamental. Potenial invasive alien species in the British isles.	30 cm.	Tel Afar, Ninaua, Assyria	Desert nearby the Tigris river, with some other reptant and crass plants (Tribulus terrestris, Apteranthes europaea).	298 m.a.s.l.	15/07/1995	36.881382N	42.492363E	https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQJhvrsQWmyOJFSJFMhPdIXuhjAONN6bXWmlQ&usqp=CAU