labeleR: Automate the Production of Custom Labels, Badges, Certificates, and Other Documents

Create custom labels, badges, certificates and other documents. Automate the production of potentially large numbers of herbarium and collection labels, accreditation badges, attendance and participation certificates, etc, and deliver them automatically. Documents are generated in PDF format, which requires a working installation of 'LaTeX', such as 'TinyTeX'.

See also

Author

Maintainer: Ignacio Ramos-Gutierrez ig.ramosgutierrez@gmail.com (ORCID) [copyright holder]

Authors:

Julia G. de Aledo juliagdealedo@gmail.com (ORCID) [copyright holder]

Jimena Mateo-Martín jimena.mateo.martin@gmail.com (ORCID) [copyright holder]

Francisco Rodriguez-Sanchez f.rodriguez.sanc@gmail.com (ORCID) [copyright holder]

Other contributors:

Giorgio Comai [contributor]